Cleaned up RabbitMQ queue flow
