+++ src/Smiosoft.PASS.RabbitMQ/Queue/QueueSubscriber.cs
@@ -5,7 +5,7 @@ using RabbitMQ.Client;
 using RabbitMQ.Client.Events;
 using Smiosoft.PASS.Extensions;
 
-namespace Smiosoft.PASS.RabbitMQ.Publisher
+namespace Smiosoft.PASS.RabbitMQ.Queue
 {
 	public abstract class QueueSubscriber<TMessage> : IQueueSubscriber<TMessage>, IDisposable
 		where TMessage : class
@@ -16,29 +16,22 @@ namespace Smiosoft.PASS.RabbitMQ.Publisher
 		protected IConnection Connection { get; }
 		protected IModel Channel { get; }
 		protected string QueueName { get; }
-		protected string RoutingKey { get; }
 
-		protected QueueSubscriber(IConnectionFactory factory, string queueName, string routingKey)
+		protected QueueSubscriber(IConnectionFactory factory, string queueName)
 		{
 			if (string.IsNullOrWhiteSpace(queueName))
 			{
 				throw new ArgumentNullException(nameof(queueName));
 			}
 
-			if (string.IsNullOrWhiteSpace(routingKey))
-			{
-				throw new ArgumentNullException(nameof(routingKey));
-			}
-
 			Factory = factory ?? throw new ArgumentNullException(nameof(factory));
 			Connection = Factory.CreateConnection();
 			Channel = Connection.CreateModel();
 			QueueName = queueName;
-			RoutingKey = routingKey;
 		}
 
-		protected QueueSubscriber(string hostName, string queueName, string routingKey)
-			: this(new ConnectionFactory() { HostName = hostName }, queueName, routingKey)
+		protected QueueSubscriber(string hostName, string queueName)
+			: this(new ConnectionFactory() { HostName = hostName }, queueName)
 		{ }
 
 		public abstract Task OnMessageRecievedAsync(TMessage message, CancellationToken cancellationToken);
@@ -50,19 +43,16 @@ namespace Smiosoft.PASS.RabbitMQ.Publisher
 
 		public virtual void Register()
 		{
-			Channel.QueueDeclare(
-				queue: QueueName,
-				durable: false,
-				exclusive: false,
-				autoDelete: false,
-				arguments: null);
+			Channel.QueueDeclare(queue: QueueName, durable: true, exclusive: false, autoDelete: false, arguments: null);
+			Channel.BasicQos(prefetchSize: 0, prefetchCount: 1, global: false);
 
 			var consumer = new EventingBasicConsumer(Channel);
-			consumer.Received += async (model, args) =>
+			consumer.Received += async (sender, args) =>
 			{
 				try
 				{
 					await OnMessageRecievedAsync(args.Body.ToArray().Deserialise<TMessage>(), CancellationToken.None);
+					Channel.BasicAck(deliveryTag: args.DeliveryTag, multiple: false);
 				}
 				catch (Exception exception)
 				{
@@ -70,10 +60,7 @@ namespace Smiosoft.PASS.RabbitMQ.Publisher
 				}
 			};
 
-			Channel.BasicConsume(
-				queue: QueueName,
-				autoAck: true,
-				consumer: consumer);
+			Channel.BasicConsume(queue: QueueName, autoAck: false, consumer: consumer);
 		}
 
 		public void Dispose()
